requirements.txt
# Simple RAG Pipeline - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
pyyaml>=6.0.1          # Configuration file parsing
requests>=2.31.0       # HTTP requests for sitemap fetching

# Web crawling
crawl4ai>=0.7.0        # AI-powered web crawler (async version)
playwright>=1.40.0     # Browser automation (required by crawl4ai 0.7+)

# Text processing and embeddings
sentence-transformers>=2.2.0  # For Stage 4: generating embeddings
numpy>=1.24.0                 # For Stage 4: vector operations
torch>=2.0.0                  # Required by sentence-transformers

# Optional: Progress bars and logging
# tqdm>=4.65.0                  # Progress bars for long operations
